requirements.txt
pandas
numpy
scikit-learn
factor-analyzer
semopy
rpy2
jinja2
plotly
networkx
mokkenpy  # if exists, or remove
pyirt
pingouinstreamlit
